WEB-3460: completed full creation round trip
