+++ src/ai/ai.service.ts
@@ -1,14 +1,26 @@
 import { ChatOpenAI } from '@langchain/openai'
 import { ChatTogetherAI } from '@langchain/community/chat_models/togetherai'
-import { Injectable, InternalServerErrorException } from '@nestjs/common'
+import { Injectable, Logger } from '@nestjs/common'
 import { AiChatMessage } from './ai.types'
+import { SlackService } from 'src/shared/services/slack.service'
+import { Prisma, User } from '@prisma/client'
+import { getFullName } from 'src/users/util/users.util'
+import {
+  CampaignDetails,
+  CampaignAiContent,
+} from 'src/campaigns/campaigns.types'
+import { againstToStr, positionsToStr, replaceAll } from './util/aiContent.util'
 
-const TOGETHER_AI_KEY = process.env.TOGETHER_AI_KEY //sails.config.custom.togetherAi || sails.config.togetherAi;
-const OPEN_AI_KEY = process.env.OPEN_AI_KEY // sails.config.custom.openAi || sails.config.openAi;
-const AI_MODELS = process.env.AI_MODELS || '' //sails.config.custom.aiModels || sails.config.aiModels || '';
+const TOGETHER_AI_KEY = process.env.TOGETHER_AI_KEY
+const OPEN_AI_KEY = process.env.OPEN_AI_KEY
+const AI_MODELS = process.env.AI_MODELS || ''
 
 @Injectable()
 export class AiService {
+  private readonly logger = new Logger(AiService.name)
+
+  constructor(private slack: SlackService) {}
+
   async llmChatCompletion(
     messages: AiChatMessage[],
     maxTokens: number = 500,
@@ -17,15 +29,13 @@ export class AiService {
   ) {
     const models = AI_MODELS.split(',')
     if (models.length === 0) {
-      // await sails.helpers.slack.slackHelper(
-      //   {
-      //     title: 'Error',
-      //     body: `AI Models are not configured. Please specify AI models.`,
-      //   },
-      //   'dev',
-      // )
-
-      throw new InternalServerErrorException('AI Models are not configured.')
+      await this.slack.slackHelper(
+        {
+          title: 'Error',
+          body: `AI Models are not configured. Please specify AI models.`,
+        },
+        'dev',
+      )
     }
 
     const aiOptions = {
@@ -37,6 +47,7 @@ export class AiService {
 
     let firstModel
     let fallbackModel
+
     for (const model of models) {
       if (model.includes('gpt')) {
         if (!firstModel) {
@@ -83,15 +94,10 @@ export class AiService {
     try {
       completion = await modelWithFallback.invoke(messages)
     } catch (error: any) {
-      console.log('Error in utils/ai/llmChatCompletion', error)
-      console.log('error response', error?.response)
+      this.logger.error('Error in utils/ai/llmChatCompletion', error)
+      this.logger.error('error response', error?.response)
 
-      // await sails.helpers.slack.errorLoggerHelper(
-      //   'Error in AI completion (raw)',
-      //   error,
-      // )
-
-      throw new InternalServerErrorException('LLM chat completion failed')
+      await this.slack.errorLoggerHelper('Error in AI completion (raw)', error)
     }
 
     if (completion && completion?.content) {
@@ -110,6 +116,353 @@ export class AiService {
         content: '',
         tokens: 0,
       }
+    }
+  }
+
+  /** function to replace placeholder tokens in ai content prompt */
+  async promptReplace(
+    prompt: string,
+    campaign: Prisma.CampaignGetPayload<{
+      include: {
+        pathToVictory: true
+        campaignPositions: {
+          include: {
+            topIssue: true
+            position: true
+          }
+        }
+        campaignUpdateHistory: true
+        user: true
+      }
+    }>,
+  ) {
+    try {
+      let newPrompt = prompt
+
+      const campaignPositions = campaign.campaignPositions
+      const user = campaign.user as User
+
+      const name = getFullName(user)
+      const details = (campaign.details || {}) as CampaignDetails
+
+      const positionsStr = positionsToStr(
+        campaignPositions,
+        details.customIssues,
+      )
+      let party =
+        details.party === 'Other' ? details.otherParty : details?.party
+      if (party === 'Independent') {
+        party = 'Independent / non-partisan'
+      }
+      const office =
+        details.office === 'Other' ? details.otherOffice : details?.office
+
+      const replaceArr = [
+        {
+          find: 'name',
+          replace: name,
+        },
+        {
+          find: 'zip',
+          replace: details.zip,
+        },
+        {
+          find: 'website',
+          replace: details.website,
+        },
+        {
+          find: 'party',
+          replace: party,
+        },
+        {
+          find: 'state',
+          replace: details.state,
+        },
+        {
+          find: 'primaryElectionDate',
+          replace: details.primaryElectionDate,
+        },
+        {
+          find: 'district',
+          replace: details.district,
+        },
+        {
+          find: 'office',
+          replace: `${office}${
+            details.district ? ` in ${details.district}` : ''
+          }`,
+        },
+        {
+          find: 'positions',
+          replace: positionsStr,
+        },
+        {
+          find: 'pastExperience',
+          replace:
+            typeof details.pastExperience === 'string'
+              ? details.pastExperience
+              : JSON.stringify(details.pastExperience || {}),
+        },
+        {
+          find: 'occupation',
+          replace: details.occupation,
+        },
+        {
+          find: 'funFact',
+          replace: details.funFact,
+        },
+        {
+          find: 'campaignCommittee',
+          replace: details.campaignCommittee || 'unknown',
+        },
+      ]
+      const againstStr = againstToStr(details.runningAgainst)
+      replaceArr.push(
+        {
+          find: 'runningAgainst',
+          replace: againstStr,
+        },
+        {
+          find: 'electionDate',
+          replace: details.electionDate,
+        },
+        {
+          find: 'statementName',
+          replace: details.statementName,
+        },
+      )
+
+      const pathToVictory = campaign.pathToVictory
+
+      if (pathToVictory) {
+        const {
+          projectedTurnout,
+          winNumber,
+          republicans,
+          democrats,
+          indies,
+          averageTurnout,
+          allAvailVoters,
+          availVotersTo35,
+          women,
+          men,
+          africanAmerican,
+          white,
+          asian,
+          hispanic,
+          voteGoal,
+          voterProjection,
+          totalRegisteredVoters,
+          budgetLow,
+          budgetHigh,
+        } = pathToVictory.data as Record<string, string | number> // TODO: better type here!!
+        replaceArr.push(
+          {
+            find: 'pathToVictory',
+            replace: JSON.stringify(pathToVictory.data),
+          },
+          {
+            find: 'projectedTurnout',
+            replace: projectedTurnout,
+          },
+          {
+            find: 'totalRegisteredVoters',
+            replace: totalRegisteredVoters,
+          },
+          {
+            find: 'winNumber',
+            replace: winNumber,
+          },
+          {
+            find: 'republicans',
+            replace: republicans,
+          },
+          {
+            find: 'democrats',
+            replace: democrats,
+          },
+          {
+            find: 'indies',
+            replace: indies,
+          },
+          {
+            find: 'averageTurnout',
+            replace: averageTurnout,
+          },
+          {
+            find: 'allAvailVoters',
+            replace: allAvailVoters,
+          },
+          {
+            find: 'availVotersTo35',
+            replace: availVotersTo35,
+          },
+          {
+            find: 'women',
+            replace: women,
+          },
+          {
+            find: 'men',
+            replace: men,
+          },
+          {
+            find: 'africanAmerican',
+            replace: africanAmerican,
+          },
+          {
+            find: 'white',
+            replace: white,
+          },
+          {
+            find: 'asian',
+            replace: asian,
+          },
+          {
+            find: 'hispanic',
+            replace: hispanic,
+          },
+          {
+            find: 'voteGoal',
+            replace: voteGoal,
+          },
+          {
+            find: 'voterProjection',
+            replace: voterProjection,
+          },
+          {
+            find: 'budgetLow',
+            replace: budgetLow,
+          },
+          {
+            find: 'budgetHigh',
+            replace: budgetHigh,
+          },
+        )
+      }
+
+      if (newPrompt.includes('[[updateHistory]]')) {
+        const updateHistoryObjects = campaign.campaignUpdateHistory
+
+        const twoWeeksAgo = new Date()
+        const thisWeek = new Date()
+        thisWeek.setDate(thisWeek.getDate() - 7)
+        twoWeeksAgo.setDate(twoWeeksAgo.getDate() - 14)
+
+        const updateHistory = {
+          allTime: {
+            total: 0,
+            doorKnocking: 0,
+            digitalAds: 0,
+            calls: 0,
+            yardSigns: 0,
+            events: 0,
+            text: 0,
+            directMail: 0,
+          },
+          thisWeek: {
+            total: 0,
+            doorKnocking: 0,
+            digitalAds: 0,
+            calls: 0,
+            yardSigns: 0,
+            events: 0,
+            text: 0,
+            directMail: 0,
+          },
+          lastWeek: {
+            total: 0,
+            doorKnocking: 0,
+            digitalAds: 0,
+            calls: 0,
+            yardSigns: 0,
+            events: 0,
+            text: 0,
+            directMail: 0,
+          },
+        }
+
+        if (updateHistoryObjects) {
+          for (const update of updateHistoryObjects) {
+            updateHistory.allTime[update.type] += update.quantity
+            updateHistory.allTime.total += update.quantity
+            if (update.createdAt > thisWeek) {
+              updateHistory.thisWeek[update.type] += update.quantity
+              updateHistory.thisWeek.total += update.quantity
+            }
+            if (update.createdAt > twoWeeksAgo && update.createdAt < thisWeek) {
+              updateHistory.lastWeek[update.type] += update.quantity
+              updateHistory.lastWeek.total += update.quantity
+            }
+          }
+        }
+        replaceArr.push({
+          find: 'updateHistory',
+          replace: JSON.stringify(updateHistory),
+        })
+      }
+
+      if (campaign.aiContent) {
+        const {
+          aboutMe,
+          communicationStrategy,
+          messageBox,
+          mobilizing,
+          policyPlatform,
+          slogan,
+          why,
+        } = campaign.aiContent as CampaignAiContent
+        replaceArr.push(
+          {
+            find: 'slogan',
+            replace: slogan?.content,
+          },
+          {
+            find: 'why',
+            replace: why?.content,
+          },
+          {
+            find: 'about',
+            replace: aboutMe?.content,
+          },
+          {
+            find: 'myPolicies',
+            replace: policyPlatform?.content,
+          },
+          {
+            find: 'commStart',
+            replace: communicationStrategy?.content,
+          },
+          {
+            find: 'mobilizing',
+            replace: mobilizing?.content,
+          },
+          {
+            find: 'positioning',
+            replace: messageBox?.content,
+          },
+        )
+      }
+
+      replaceArr.forEach((item) => {
+        try {
+          newPrompt = replaceAll(
+            newPrompt,
+            item.find,
+            item.replace ? item.replace.toString().trim() : '',
+          )
+        } catch (e) {
+          this.logger.error('error at prompt replace', e)
+        }
+      })
+
+      newPrompt += `\n
+        
+      `
+
+      return newPrompt
+    } catch (e) {
+      this.logger.error('Error in helpers/ai/promptReplace', e)
+      return ''
     }
   }
 }
