+++ src/ai/util/aiContent.util.ts
