+++ src/campaigns/ai/campaignsAi.controller.ts
@@ -4,6 +4,8 @@ import {
   Delete,
   HttpCode,
   HttpStatus,
+  InternalServerErrorException,
+  NotFoundException,
   Param,
   Post,
   Put,
@@ -17,7 +19,8 @@ import { ZodValidationPipe } from 'nestjs-zod'
 import { User } from '@prisma/client'
 import { CreateAiContentSchema } from './schemas/CreateAiContent.schema'
 import { FastifyReply } from 'fastify'
-import { CampaignsService } from '../campaigns.service'
+import { CampaignsService } from '../services/campaigns.service'
+import { CampaignAiContent } from '../campaigns.types'
 
 @Controller('campaigns/ai')
 @UsePipes(ZodValidationPipe)
@@ -34,32 +37,63 @@ export class CampaignsAiController {
     @Body() body: CreateAiContentSchema,
   ) {
     const campaign = await this.loadCampaign(userId)
-    const result = await this.campaignsAiService.createContent(campaign, body)
 
-    if (result.created) {
-      res.statusCode = HttpStatus.CREATED
-    } else {
-      res.statusCode = HttpStatus.OK
+    try {
+      const result = await this.campaignsAiService.createContent(campaign, body)
+
+      if (result.created) {
+        res.statusCode = HttpStatus.CREATED
+      } else {
+        res.statusCode = HttpStatus.OK
+      }
+
+      return result
+    } catch (e) {
+      if (e instanceof Error) {
+        throw new InternalServerErrorException('Failed to create content')
+      }
+
+      throw e
     }
-
-    return result
   }
 
   @Put('rename')
   @HttpCode(HttpStatus.OK)
   async rename(
     @ReqUser() { id: userId }: User,
-    @Body() body: RenameAiContentSchema,
+    @Body() { key, name }: RenameAiContentSchema,
   ) {
     const campaign = await this.loadCampaign(userId)
-    return this.campaignsAiService.updateContentName(campaign, body)
+    const { aiContent } = campaign
+
+    if (!aiContent?.[key]) {
+      throw new NotFoundException(`Content with key: ${key} not found`)
+    }
+
+    aiContent[key]['name'] = name
+
+    return this.campaignsService.update(campaign.id, {
+      aiContent,
+    })
   }
 
   @Delete(':key')
   @HttpCode(HttpStatus.NO_CONTENT)
   async delete(@ReqUser() { id: userId }: User, @Param('key') key: string) {
     const campaign = await this.loadCampaign(userId)
-    return this.campaignsAiService.deleteContent(campaign, key)
+    const aiContent = campaign.aiContent as CampaignAiContent
+
+    if (!aiContent?.[key]) {
+      // nothing to delete
+      throw new NotFoundException('Content not found')
+    }
+
+    delete aiContent[key]
+    delete aiContent.generationStatus?.[key]
+
+    return this.campaignsService.update(campaign.id, {
+      aiContent,
+    })
   }
 
   private loadCampaign(userId: number) {
