+++ src/queue/queue.module.ts
@@ -1,9 +1,10 @@
-import { Module } from '@nestjs/common'
+import { forwardRef, Module } from '@nestjs/common'
 import { SqsModule } from '@ssut/nestjs-sqs'
 import { EnqueueService } from './enqueue.service'
 import { ConsumerService } from './consumer.service'
 import { QueueController } from './queue.controller'
 import { queueConfig } from './queue.config'
+import { CampaignsAiModule } from 'src/campaigns/ai/campaignsAi.module'
 
 @Module({
   imports: [
@@ -15,6 +16,7 @@ import { queueConfig } from './queue.config'
         },
       ],
     }),
+    forwardRef(() => CampaignsAiModule),
   ],
   controllers: [QueueController],
   providers: [ConsumerService, EnqueueService],
