+++ src/shared/shared.module.ts
@@ -1,0 +1,9 @@
+import { Global, Module } from '@nestjs/common'
+import { SlackService } from './services/slack.service'
+
+@Global()
+@Module({
+  providers: [SlackService],
+  exports: [SlackService],
+})
+export class SharedModule {}
