+++ src/shared/util/strings.util.ts
@@ -31,3 +31,8 @@ export const generateRandomString = (
   ]
     .map((b) => CHARSET[b % CHARSET.length])
     .join('')
+
+export function camelToSentence(text) {
+  const result = text.replace(/([A-Z])/g, ' $1')
+  return result.charAt(0).toUpperCase() + result.slice(1)
+}
